test(terraform): Move tests which do not require env vars into unit tests allowing for version mocking and removing another install step from our ci
