+++ .github/workflows/workflow.yml
@@ -121,12 +121,6 @@ jobs:
         if: matrix.os == 'macOS-latest'
         run: pip3 install mercurial
 
-      # Install Terraform at a fixed version
-      - name: Setup | Terraform
-        uses: volcano-coffee-company/setup-terraform@v1
-        with:
-          version: "0.12.14"
-
       # Run the ignored tests that expect the above setup
       - name: Build | Test
         run: cargo test -- -Z unstable-options --include-ignored
