+++ src/utils.rs
@@ -123,6 +123,10 @@ active boot switches: -d:release\n",
             stdout: String::from("0.6.0"),
             stderr: String::default(),
         }),
+        "terraform version" => Some(CommandOutput {
+            stdout: String::from("Terraform v0.12.14"),
+            stderr: String::default(),
+        }),
         s if s.starts_with("erl") => Some(CommandOutput {
             stdout: String::from("22.1.3"),
             stderr: String::default(),
